+++ services/app-api/src/domain-models/contractAndRates/parseDomainData.ts
@@ -1,5 +1,4 @@
 import {
-    ContractTableWithRelations,
     DraftContractRevisionTableWithRelations,
     DraftContractTableWithRelations,
 } from '../../postgres/prismaTypes'
@@ -8,12 +7,10 @@ import {
     ContractRevisionType,
     contractRevisionZodSchema,
     draftContractZodSchema,
-    contractZodSchema,
 } from './contractAndRatesZodSchema'
 import {
     draftContractToDomainModel,
     draftContractRevToDomainModel,
-    contractWithHistoryToDomainModel,
 } from '../../postgres/contractAndRates/prismaToDomainModel'
 
 function parseDraftContractRevision(
@@ -51,36 +48,4 @@ function parseDraftContract(
     return parseDraft.data
 }
 
-// parseContractWithHistory returns a ContractType with a full set of
-// ContractRevisions in reverse chronological order. Each revision is a change to this
-// Contract with submit and unlock info. Changes to the data of this contract, or changes
-// to the data or relations of associate revisions will all surface as new ContractRevisions
-function parseContractWithHistory(
-    contract: ContractTableWithRelations
-    //contractRevisions: ContractRevisionTableWithRelations[]
-): ContractType | Error {
-    const contractWithHistory = contractWithHistoryToDomainModel(contract)
-
-    if (contractWithHistory instanceof Error) {
-        console.warn(
-            `ERROR: attempting to parse prisma contract with history failed: ${contractWithHistory.message}`
-        )
-        return contractWithHistory
-    }
-
-    const parseContract = contractZodSchema.safeParse(contractWithHistory)
-
-    if (!parseContract.success) {
-        const error = `ERROR: attempting to parse prisma contract with history failed: ${parseContract.error}`
-        console.warn(error)
-        return parseContract.error
-    }
-
-    return parseContract.data
-}
-
-export {
-    parseDraftContractRevision,
-    parseDraftContract,
-    parseContractWithHistory,
-}
+export { parseDraftContractRevision, parseDraftContract }
