+++ services/app-api/src/postgres/contractAndRates/prismaToDomainModel.ts
@@ -12,20 +12,10 @@ import {
     DraftRateWithRelations,
     ContractRevisionTableWithRelations,
     ContractRevisionFormDataType,
-    ContractTableWithRelations,
 } from '../prismaTypes'
 import { UpdateInfoType } from '../../domain-models'
 import { DocumentCategoryType } from 'app-web/src/common-code/healthPlanFormDataType'
 import { RateRevisionTable } from '@prisma/client'
-
-// ContractRevisionSet is for the internal building of individual revisions
-// we convert them into ContractRevisions to return them
-interface ContractRevisionSet {
-    contractRev: ContractRevisionTableWithRelations
-    submitInfo: UpdateInfoTableWithUpdater
-    unlockInfo: UpdateInfoTableWithUpdater | undefined
-    rateRevisions: RateRevisionTable[]
-}
 
 function convertUpdateInfoToDomainModel(
     info?: UpdateInfoTableWithUpdater | null
@@ -165,24 +155,6 @@ function draftContractRevToDomainModel(
     }
 }
 
-function contractRevToDomainModel(
-    revisions: ContractRevisionSet[]
-): ContractRevisionType[] {
-    const contractRevisions = revisions.map((entry) => ({
-        id: entry.contractRev.id,
-        submitInfo: convertUpdateInfoToDomainModel(entry.submitInfo),
-        unlockInfo: entry.unlockInfo
-            ? convertUpdateInfoToDomainModel(entry.unlockInfo)
-            : undefined,
-        createdAt: entry.contractRev.createdAt,
-        updatedAt: entry.contractRev.updatedAt,
-        formData: contractFormDataToDomainModel(entry.contractRev),
-        rateRevisions: ratesRevisionsToDomainModel(entry.rateRevisions),
-    }))
-
-    return contractRevisions
-}
-
 function draftContractToDomainModel(
     contract: DraftContractTableWithRelations
 ): ContractType {
@@ -199,96 +171,12 @@ function draftContractToDomainModel(
     }
 }
 
-// contractWithHistoryToDomainModel constructs a history for this particular contract including changes to all of its
-// revisions and all related rate revisions, including added and removed rates
-function contractWithHistoryToDomainModel(
-    contract: ContractTableWithRelations
-): ContractType | Error {
-    // We iterate through each contract revision in order, adding it as a revision in the history
-    // then iterate through each of its rates, constructing a history of any rates that changed
-    // between contract revision updates
-    const allRevisionSets: ContractRevisionSet[] = []
-    const contractRevisions = contract.revisions
-    for (const contractRev of contractRevisions) {
-        // We exclude the draft from this list, use findDraftContract to get the current draft
-        if (!contractRev.submitInfo) {
-            continue
-        }
-
-        const initialEntry: ContractRevisionSet = {
-            contractRev,
-            submitInfo: contractRev.submitInfo,
-            unlockInfo: contractRev.unlockInfo || undefined,
-            rateRevisions: [],
-        }
-
-        allRevisionSets.push(initialEntry)
-
-        let lastEntry = initialEntry
-        // go through every rate revision in the join table in time order and construct a revisionSet
-        // with (or without) the new rate revision in it.
-        for (const rateRev of contractRev.rateRevisions) {
-            if (!rateRev.rateRevision.submitInfo) {
-                return new Error(
-                    'Programming Error: a contract is associated with an unsubmitted rate'
-                )
-            }
-
-            // if it's from before this contract was submitted, it's there at the beginning.
-            if (
-                rateRev.rateRevision.submitInfo.updatedAt <=
-                contractRev.submitInfo.updatedAt
-            ) {
-                if (!rateRev.isRemoval) {
-                    initialEntry.rateRevisions.push(rateRev.rateRevision)
-                }
-            } else {
-                // if after, then it's always a new entry in the list
-                let lastRates = [...lastEntry.rateRevisions]
-
-                // take out the previous rate revision this revision supersedes
-                lastRates = lastRates.filter(
-                    (r) => r.rateID !== rateRev.rateRevision.rateID
-                )
-                // an isRemoval entry indicates that this rate was removed from this contract.
-                if (!rateRev.isRemoval) {
-                    lastRates.push(rateRev.rateRevision)
-                }
-
-                const newRev: ContractRevisionSet = {
-                    contractRev,
-                    submitInfo: rateRev.rateRevision.submitInfo,
-                    unlockInfo: rateRev.rateRevision.unlockInfo || undefined,
-                    rateRevisions: lastRates,
-                }
-
-                lastEntry = newRev
-                allRevisionSets.push(newRev)
-            }
-        }
-    }
-
-    const revisions = contractRevToDomainModel(allRevisionSets).reverse()
-
-    const contractStatus = getContractStatus(contract.revisions)
-
-    return {
-        id: contract.id,
-        status: contractStatus,
-        stateCode: contract.stateCode,
-        stateNumber: contract.stateNumber,
-        revisions: revisions,
-    }
-}
-
 export {
     contractFormDataToDomainModel,
     convertUpdateInfoToDomainModel,
     draftContractRevToDomainModel,
     draftContractToDomainModel,
-    contractRevToDomainModel,
     draftRatesToDomainModel,
     ratesRevisionsToDomainModel,
-    contractWithHistoryToDomainModel,
     getContractStatus,
 }
